Add blur effect on Contact button
